Refactors cheat config to unified useConfig hook
Replaces legacy cheat flag access with a new useConfig hook for feature toggles and environment-driven flags. Unifies configuration logic, enables support for query param overrides, and cleans up related code across components and hooks. Removes redundant helpers and centralizes allChains logic for multi-chain queries. Improves maintainability and flexibility of feature flag management.
