+++ apps/web/app/(app)/gardens/page.tsx
@@ -19,7 +19,7 @@ import {
 } from "@/assets";
 import { Button, Communities } from "@/components";
 import { LightCommunity } from "@/components/Communities";
-import { useCheat } from "@/hooks/useCheat";
+import { useConfig } from "@/hooks/useCheat";
 import { useDisableButtons } from "@/hooks/useDisableButtons";
 import { useSubgraphQueryMultiChain } from "@/hooks/useSubgraphQueryMultiChain";
 import { useTheme } from "@/providers/ThemeProvider";
@@ -113,7 +113,7 @@ export default function GardensPage() {
     undefined,
   );
 
-  const showArchived = useCheat("showArchived");
+  const showArchived = useConfig("showArchived");
 
   useEffect(() => {
     getProtopiansOwners()
