+++ apps/web/hooks/useSubgraphQuery.ts
@@ -3,7 +3,7 @@ import { AnyVariables, DocumentInput, OperationContext } from "@urql/next";
 import { isEqual } from "lodash-es";
 import { toast } from "react-toastify";
 import { useChainIdFromPath } from "./useChainIdFromPath";
-import { useCheat } from "./useCheat";
+import { useConfig } from "./useCheat";
 import { useIsMounted } from "./useIsMounted";
 import { LoadingToast } from "@/components";
 import { getConfigByChain } from "@/configs/chains";
@@ -70,7 +70,7 @@ export function useSubgraphQuery<
   const latestResponse = useRef({ variables, response });
   const subscritionId = useRef<SubscriptionId>();
   const fetchingRef = useRef(false);
-  const skipPublished = useCheat("skipPublished");
+  const skipPublished = useConfig("skipPublished");
 
   useEffect(() => {
     latestResponse.current.response = response; // Update ref on every response change
@@ -137,9 +137,7 @@ export function useSubgraphQuery<
 
     let res;
     try {
-      const shouldSkipPublished =
-        skipPublished || process.env.NEXT_PUBLIC_SKIP_PUBLISHED === "true";
-      res = await urqlQuery(shouldSkipPublished);
+      res = await urqlQuery(skipPublished);
       if (!res.data && res.error) {
         throw res.error;
       }
