+++ apps/web/hooks/useSubgraphQueryMultiChain.ts
@@ -7,7 +7,7 @@ import {
 } from "@urql/next";
 import { debounce, isEqual } from "lodash-es";
 import { toast } from "react-toastify";
-import { getCheat, useCheat } from "./useCheat";
+import { useConfig } from "./useCheat";
 import { useIsMounted } from "./useIsMounted";
 import { HTTP_CODES } from "@/app/api/utils";
 import { LoadingToast } from "@/components";
@@ -25,21 +25,6 @@ import {
 import { initUrqlClient } from "@/providers/urql";
 import { ChainId } from "@/types";
 import { delayAsync } from "@/utils/delayAsync";
-
-let isQueryAllChains = false;
-try {
-  isQueryAllChains = getCheat("queryAllChains");
-} catch (error) {
-  // ignore when not browser side
-}
-
-export const allChains: ChainId[] = Object.entries(chainConfigMap)
-  .filter(
-    ([_, chainConfig]) =>
-      isQueryAllChains ||
-      (isProd ? !chainConfig.isTestnet : !!chainConfig.isTestnet),
-  )
-  .map(([chainId]) => Number(chainId));
 
 const pendingRefreshToastId = "pending-refresh";
 
@@ -72,7 +57,16 @@ export function useSubgraphQueryMultiChain<
   const errorsMap = useRef(new Map<ChainId, CombinedError>());
   const subscritionId = useRef<SubscriptionId>();
   const fetchingRef = useRef(false);
-  const skipPublished = useCheat("skipPublished");
+  const skipPublished = useConfig("skipPublished");
+  const isQueryAllChains = useConfig("queryAllChains");
+
+  const allChains = Object.entries(chainConfigMap)
+    .filter(
+      ([_, chainConfig]) =>
+        isQueryAllChains ||
+        (isProd ? !chainConfig.isTestnet : !!chainConfig.isTestnet),
+    )
+    .map(([chainId]) => Number(chainId));
 
   useEffect(() => {
     if (!enabled) return;
@@ -167,10 +161,7 @@ export function useSubgraphQueryMultiChain<
 
               let res;
               try {
-                const shouldSkipPublished =
-                  skipPublished ||
-                  process.env.NEXT_PUBLIC_SKIP_PUBLISHED === "true";
-                res = await fetchQuery(shouldSkipPublished);
+                res = await fetchQuery(skipPublished);
                 if (res.data == null && res.error) {
                   throw res.error;
                 }
