+++ turbo.json
@@ -54,11 +54,9 @@
         "RPC_URL_MATIC",
         "RPC_URL_BASE",
         "CRON_SECRET",
-        "NEXT_PUBLIC_SKIP_PUBLISHED",
         "CI",
         "NEXT_PUBLIC_SUBGRAPH_KEY",
-        "NEXT_PUBLIC_ALCHEMY_KEY",
-        "NEXT_PUBLIC_SKIP_PUBLISHED"
+        "NEXT_PUBLIC_ALCHEMY_KEY"
       ]
     },
     "test": {
